+++ app/src/main/java/com/intellistart/interviewplanning/slot/infrastructure/adapters/nats/slot/UpdateSlotNatsController.kt
@@ -2,8 +2,8 @@ package com.intellistart.interviewplanning.slot.infrastructure.adapters.nats.slo
 
 import com.google.protobuf.Parser
 import com.intellistart.interviewplanning.NatsSubject
-import com.intellistart.interviewplanning.appltication.port.PeriodOperationsInPort
 import com.intellistart.interviewplanning.exceptions.UserException
+import com.intellistart.interviewplanning.port.PeriodOperationsInPort
 import com.intellistart.interviewplanning.request.slot.update.proto.UpdateSlotRequest
 import com.intellistart.interviewplanning.request.slot.update.proto.UpdateSlotResponse
 import com.intellistart.interviewplanning.slot.domain.exception.SlotException
